修改协议ID类型为int
修改协议ID类型为int

+++ Assets/GameMain/Scripts/LiteNetLib.Test/Client/OutsideNetManager.cs
@@ -107,7 +107,7 @@ namespace GameMain
             }
             _cachedWriter.Reset();
             _cachedWriter.Put((byte)PacketType.Serialized);
-            ulong protoId = ProtoIDGetter.TryGetId<T>();
+            int protoId = ProtoIDGetter.TryGetId<T>();
             _cachedWriter.Put(protoId);
             packet.Serialize(_cachedWriter);
             _server.Send(_cachedWriter);
